refactor project src dir structure
This will help with better organization/isolation of code-base for various tools and improves readability/consistency.
