+++ src/extension.ts
@@ -1,10 +1,11 @@
 import * as vscode from "vscode";
-import { adb, apktool } from "./tools";
-import { outputChannel } from "./common";
-import { updateTools } from "./downloader";
+import { outputChannel } from "./data/constants";
+import { updateTools } from "./utils/downloader";
 import { UI } from "./interface";
-import { applyMitmPatches } from "./mitm-patches";
-import { Quark } from "./quark-tools";
+import { applyMitmPatches } from "./tools/apk-mitm";
+import { Quark } from "./tools/quark-engine";
+import { adb } from "./tools/adb";
+import { apktool } from "./tools/apktool";
 
 export function activate(context: vscode.ExtensionContext): void {
     console.log("Activated apklab extension!");
