+++ src/tools/apk-mitm.ts
@@ -1,7 +1,7 @@
 import * as path from "path";
-import * as apkMitm from "apk-mitm";
+import { applyPatches, observeListr } from "apk-mitm";
 import * as vscode from "vscode";
-import { outputChannel } from "./common";
+import { outputChannel } from "../data/constants";
 
 // Defined in webpack config
 declare const APK_MITM_VERSION: string;
@@ -25,9 +25,9 @@ export async function applyMitmPatches(apktoolYmlPath: string): Promise<void> {
 
         const projectDir = path.dirname(apktoolYmlPath);
 
-        await apkMitm
-            .observeListr(apkMitm.applyPatches(projectDir))
-            .forEach((line) => outputChannel.appendLine(line));
+        await observeListr(applyPatches(projectDir)).forEach((line) =>
+            outputChannel.appendLine(line)
+        );
 
         outputChannel.appendLine("\nSuccessfully applied MITM patches!");
         vscode.window.showInformationMessage(
